requirements.txt
playwright>=1.58.0
python-telegram-bot>=21.6
httpx>=0.28.1
python-dotenv>=1.2.2
tenacity>=9.1.2
